Part of making MSVC happy

+++ src/impl/kscope_prng.h
@@ -112,7 +112,7 @@ namespace ithare {
 			v ^= v << 20; v ^= v >> 41; v ^= v << 5;
 			return v;
 		}
-		constexpr uint32_t kscope_random(uint64_t seed, int32_t modifier, uint32_t maxn) {
+		constexpr size_t kscope_random(uint64_t seed, int32_t modifier, size_t maxn) {
 			//for maxn < 1M, bias is limited to <0.1% - more than enough for our purposes
 			assert(maxn > 0);
 			assert(maxn < 1'048'576);//limiting bits to 20 and bias to <0.1%
